Made Config an EventEmitter, working on Admins
Config now emits events, will be useful in Admins so the object can refresh when new admins are added/deleted.  Later, this will be really useful when new channels/guilds are added for logging.

Admins is still in progress, but it will now create a Map with all defined admins and their Discord.User objects and send a message to each admin when the bot logs in.  DMs from admins are also being logged to the console right now as I continue to build this object out.