Trying to figure out a command structure
Most of the changes have come in Admins, I can't figure out the best way to design a useful command structure.  I want one that will accept commands like this:

> list
Would you like to list admins, channels, or guilds?
> admins
[list of admins]

but it would also work like this:

> list admins
[list of admins]

sounds easy, but implementing it and making it as modular as possible in the backend to make adding/modifying commands easily is giving me a fucking hard ass time.