Lots of changes! So I was out of the country for about two weeks and forgot to do a commit.  Then I came back and worked on the project and forgot to do a commit.  Well, here's my stinkin' commit!
Admins is likely going to be deprecated, I've come up with a new solution that should finally be robust enough to do what I want it to do.  I'll be adding a few more files in the next commit.

Also, the project finally has a name, ponardisc.  it's dumb, but hey, it's better than ponyarchivediscord, right?

Modified README and other files to reflect the new name.